Move templated config to contrib
